backend/pyproject.toml
[project]
name = "bearwithme-backend"
version = "0.1.0"
description = "BearWithMe mental wellness companion — Python backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",

    # MongoDB
    "motor>=3.7",
    "beanie>=1.29",

    # Auth
    "python-jose[cryptography]>=3.3",
    "google-auth>=2.38",
    "httpx>=0.28",

    # AI / LangChain
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.4",
    "langchain-mongodb>=0.5",

    # Production
    "pydantic-settings>=2.8",
    "slowapi>=0.1",
    "structlog>=25.1",
    "sse-starlette>=2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]
